requirements.txt
matplotlib>=3.7
numpy>=1.23
Pillow>=10.0
PyYAML>=6.0
scikit-learn>=1.3
seaborn>=0.13
tensorflow>=2.10
pytest>=7.4
